+++ match/src/Piipan.Match/Piipan.Match.Api/Models/OrchMatchResponse.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace Piipan.Match.Core.Models
+namespace Piipan.Match.Api.Models
 {
     /// <summary>
     /// Represents the top-level response body for a successful API response
