+++ match/src/Piipan.Match/Piipan.Match.Core/Parsers/OrchMatchRequestParser.cs
@@ -2,7 +2,7 @@ using System;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Piipan.Match.Core.Models;
+using Piipan.Match.Api.Models;
 using Newtonsoft.Json;
 using FluentValidation;
