+++ match/src/Piipan.Match/Piipan.Match.Core/Resolvers/MatchResolver.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using System.Threading.Tasks;
+using Piipan.Match.Api.Models;
 using Piipan.Match.Core.Extensions;
-using Piipan.Match.Core.Models;
 using Piipan.Participants.Api;
 using FluentValidation;
