+++ match/tests/Piipan.Match.Core.Tests/Validators/OrchMatchRequestValidatorTests.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using Piipan.Match.Core.Models;
+using Piipan.Match.Api.Models;
 using Piipan.Match.Core.Validators;
 using FluentValidation.TestHelper;
 using Xunit;
