+++ match/tests/Piipan.Match.Func.Api.IntegrationTests/Piipan.Match.Func.Api.IntegrationTests.csproj
@@ -22,6 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\Piipan.Match\Piipan.Match.Api\Piipan.Match.Api.csproj" />
     <ProjectReference Include="..\..\src\Piipan.Match\Piipan.Match.Func.Api\Piipan.Match.Func.Api.csproj" />
     <ProjectReference Include="..\..\..\participants\src\Piipan.Participants\Piipan.Participants.Api\Piipan.Participants.Api.csproj" />
     <ProjectReference Include="..\..\..\participants\src\Piipan.Participants\Piipan.Participants.Core\Piipan.Participants.Core.csproj" />
